Define ConstraintFactory as a test service
